+++ src/BuddyClimb/Patches/CharacterInteractiblePatch.cs
@@ -34,9 +34,19 @@ internal static class CharacterInteractiblePatch
 
         if (CanBeClimbed(__instance.character) && CanClimb(interactor))
         {
-            Photon.Realtime.Player passOutRpcTarget = GetPassOutRpcTarget(__instance.character);
-            interactor.photonView.RPC("RPCA_PassOut", passOutRpcTarget);
-            __instance.character.refs.carriying.StartCarry(interactor);
+            if (DebugPlayerSpawner.IsDebugSpawnedPlayer(__instance.character))
+            {
+                __instance.character.photonView.RPC(
+                    nameof(CharacterCarrying.RPCA_StartCarry),
+                    RpcTarget.All,
+                    interactor.photonView);
+            }
+            else
+            {
+                Photon.Realtime.Player passOutRpcTarget = GetPassOutRpcTarget(__instance.character);
+                interactor.photonView.RPC("RPCA_PassOut", passOutRpcTarget);
+                __instance.character.refs.carriying.StartCarry(interactor);
+            }
         }
     }
 
@@ -68,59 +78,6 @@ internal static class CharacterInteractiblePatch
         {
             __result = true;
         }
-    }
-
-    [HarmonyPatch(typeof(CharacterCarrying), nameof(CharacterCarrying.RPCA_Drop))]
-    [HarmonyPrefix]
-    private static bool RPCA_DropPrefix(PhotonView targetView)
-    {
-        if (targetView == null || !targetView.IsMine)
-        {
-            return true;
-        }
-
-        Character character = targetView.GetComponent<Character>();
-        Character? carrier = character?.data.carrier;
-        if (character != null && carrier != null && !character.data.fullyPassedOut)
-        {
-            targetView.RPC("RPCA_UnPassOut", carrier.photonView.Owner);
-        }
-
-        return true;
-    }
-
-    [HarmonyPatch(typeof(CharacterCarrying), nameof(CharacterCarrying.RPCA_StartCarry))]
-    [HarmonyPrefix]
-    private static bool RPCA_StartCarryPrefix(CharacterCarrying __instance, PhotonView targetView)
-    {
-        if (!DebugPlayerSpawner.IsDebugSpawnedPlayer(__instance.character))
-        {
-            return true;
-        }
-
-        Character carriedCharacter = targetView.GetComponent<Character>();
-        if (carriedCharacter == null)
-        {
-            return false;
-        }
-
-        if (__instance.character.data.carriedPlayer != null)
-        {
-            __instance.Drop(__instance.character.data.carriedPlayer);
-            return false;
-        }
-
-        carriedCharacter.refs.carriying.ToggleCarryPhysics(true);
-        carriedCharacter.data.isCarried = true;
-        __instance.character.data.carriedPlayer = carriedCharacter;
-        carriedCharacter.data.carrier = __instance.character;
-
-        foreach (Character playerCharacter in PlayerHandler.GetAllPlayerCharacters())
-        {
-            playerCharacter.refs.afflictions.UpdateWeight();
-        }
-
-        return false;
     }
 
     private static bool CanBeClimbed(Character character)